Add a NullablePoint class and use it instead of Point<double>
charts_common expects to be able use null coordinate values, but when
null-safety is enabled, this will no longer possible with
`dart:math`'s `Point` class.

Add a `NullablePoint` class to use as a substitute.  This might not
be the ideal long-term solution, but it should preserve existing
behavior for now.  I also have replaced only the sites that clearly
depended on null coordinates; it is quite possible that there are
sites not covered by automated tests that still need to switch from
`Point<double>` to `NullablePoint`.

PiperOrigin-RevId: 367165943